Deprecate conda.plan
