+++ conda/misc.py
@@ -7,10 +7,12 @@ import re
 import shutil
 import sys
 from collections import defaultdict
+from logging import getLogger
 from os.path import abspath, dirname, exists, isdir, isfile, join, relpath
 
 from .base.context import context
 from .common.compat import on_mac, on_win, open
+from .common.io import dashlist
 from .common.path import expand
 from .common.url import is_url, join_url, path_to_url
 from .core.index import get_index
@@ -26,9 +28,10 @@ from .exceptions import (
 )
 from .gateways.disk.delete import rm_rf
 from .gateways.disk.link import islink, readlink, symlink
-from .models.match_spec import MatchSpec
+from .models.match_spec import ChannelMatch, MatchSpec
 from .models.prefix_graph import PrefixGraph
-from .plan import _get_best_prec_match
+
+log = getLogger(__name__)
 
 
 def conda_installed_files(prefix, exclude_self_build=False):
@@ -341,3 +344,18 @@ def clone_env(prefix1, prefix2, verbose=True, quiet=False, index_args=None):
         index_args=index_args,
     )
     return actions, untracked_files
+
+
+def _get_best_prec_match(precs):
+    assert precs
+    for chn in context.channels:
+        channel_matcher = ChannelMatch(chn)
+        prec_matches = tuple(
+            prec for prec in precs if channel_matcher.match(prec.channel.name)
+        )
+        if prec_matches:
+            break
+    else:
+        prec_matches = precs
+    log.warn("Multiple packages found:%s", dashlist(prec_matches))
+    return prec_matches[0]
