+++ conda/plan.py
@@ -18,31 +18,27 @@ from boltons.setutils import IndexedSet
 
 from .base.constants import DEFAULTS_CHANNEL_NAME, UNKNOWN_CHANNEL
 from .base.context import context, reset_context
+from .cli.install import revert_actions as _revert_actions
 from .common.constants import TRACE
 from .common.io import dashlist, env_vars, time_recorder
 from .common.iterators import groupby_to_dict as groupby
-from .core.index import LAST_CHANNEL_URLS, _supplement_index_with_prefix
+from .core.index import LAST_CHANNEL_URLS
 from .core.link import PrefixSetup, UnlinkLinkTransaction
-from .core.solve import diff_for_unlink_link_precs
 from .deprecations import deprecated
-from .exceptions import CondaIndexError, PackagesNotFoundError
-from .history import History
 from .instructions import FETCH, LINK, SYMLINK_CONDA, UNLINK
+from .misc import _get_best_prec_match as __get_best_prec_match
 from .models.channel import Channel, prioritize_channels
 from .models.dist import Dist
 from .models.enums import LinkType
-from .models.match_spec import ChannelMatch
-from .models.prefix_graph import PrefixGraph
+from .models.match_spec import MatchSpec
 from .models.records import PackageRecord
 from .models.version import normalized_version
-from .resolve import MatchSpec
 from .utils import human_bytes
 
 log = getLogger(__name__)
 
-# TODO: Remove conda/plan.py.  This module should be almost completely deprecated now.
 
-
+@deprecated("24.9", "25.3", addendum="Unused.")
 def print_dists(dists_extras):
     fmt = "    %-27s|%17s"
     print(fmt % ("package", "build"))
@@ -54,6 +50,7 @@ def print_dists(dists_extras):
         print(line)
 
 
+@deprecated("24.9", "25.3", addendum="Unused.")
 def display_actions(
     actions, index, show_channel_urls=None, specs_to_remove=(), specs_to_add=()
 ):
@@ -265,6 +262,7 @@ def display_actions(
     print()
 
 
+@deprecated("24.9", "25.3", addendum="Unused.")
 def add_unlink(actions, dist):
     assert isinstance(dist, Dist)
     if UNLINK not in actions:
@@ -272,76 +270,35 @@ def add_unlink(actions, dist):
     actions[UNLINK].append(dist)
 
 
-# -------------------------------------------------------------------
-
-
+@deprecated("24.9", "25.3", addendum="Unused.")
 def add_defaults_to_specs(r, linked, specs, update=False, prefix=None):
     return
 
 
-def _get_best_prec_match(precs):
-    assert precs
-    for chn in context.channels:
-        channel_matcher = ChannelMatch(chn)
-        prec_matches = tuple(
-            prec for prec in precs if channel_matcher.match(prec.channel.name)
-        )
-        if prec_matches:
-            break
-    else:
-        prec_matches = precs
-    log.warn("Multiple packages found:%s", dashlist(prec_matches))
-    return prec_matches[0]
+deprecated.constant(
+    "24.9",
+    "25.3",
+    "_get_best_prec_match",
+    __get_best_prec_match,
+    addendum="Use `conda.misc._get_best_prec_match` instead.",
+)
+deprecated.constant(
+    "24.9",
+    "25.3",
+    "revert_actions",
+    _revert_actions,
+    addendum="Use `conda.cli.install.revert_actions` instead.",
+)
 
 
-def revert_actions(prefix, revision=-1, index=None):
-    # TODO: If revision raise a revision error, should always go back to a safe revision
-    h = History(prefix)
-    # TODO: need a History method to get user-requested specs for revision number
-    #       Doing a revert right now messes up user-requested spec history.
-    #       Either need to wipe out history after ``revision``, or add the correct
-    #       history information to the new entry about to be created.
-    # TODO: This is wrong!!!!!!!!!!
-    user_requested_specs = h.get_requested_specs_map().values()
-    try:
-        target_state = {
-            MatchSpec.from_dist_str(dist_str) for dist_str in h.get_state(revision)
-        }
-    except IndexError:
-        raise CondaIndexError("no such revision: %d" % revision)
-
-    _supplement_index_with_prefix(index, prefix)
-
-    not_found_in_index_specs = set()
-    link_precs = set()
-    for spec in target_state:
-        precs = tuple(prec for prec in index.values() if spec.match(prec))
-        if not precs:
-            not_found_in_index_specs.add(spec)
-        elif len(precs) > 1:
-            link_precs.add(_get_best_prec_match(precs))
-        else:
-            link_precs.add(precs[0])
-
-    if not_found_in_index_specs:
-        raise PackagesNotFoundError(not_found_in_index_specs)
-
-    final_precs = IndexedSet(PrefixGraph(link_precs).graph)  # toposort
-    unlink_precs, link_precs = diff_for_unlink_link_precs(prefix, final_precs)
-    stp = PrefixSetup(prefix, unlink_precs, link_precs, (), user_requested_specs, ())
-    txn = UnlinkLinkTransaction(stp)
-    return txn
-
-
-# ---------------------------- Backwards compat for conda-build --------------------------
-
-
+@deprecated("24.9", "25.3", addendum="Unused.")
 @time_recorder("execute_actions")
 def execute_actions(actions, index, verbose=False):  # pragma: no cover
     plan = _plan_from_actions(actions, index)
     execute_instructions(plan, index, verbose)
 
 
+@deprecated("24.9", "25.3", addendum="Unused.")
 def _plan_from_actions(actions, index):  # pragma: no cover
     from .instructions import ACTION_CODES, PREFIX, PRINT, PROGRESS, PROGRESS_COMMANDS
 
@@ -391,10 +348,10 @@ def _plan_from_actions(actions, index):  # pragma: no cover
     return plan
 
 
+@deprecated("24.9", "25.3", addendum="Unused.")
 def _inject_UNLINKLINKTRANSACTION(plan, index, prefix, axn, specs):  # pragma: no cover
     from os.path import isdir
 
-    from .core.link import PrefixSetup, UnlinkLinkTransaction
     from .core.package_cache_data import ProgressiveFetchExtract
     from .instructions import (
         LINK,
@@ -437,6 +394,7 @@ def _inject_UNLINKLINKTRANSACTION(plan, index, prefix, axn, specs):  # pragma: n
     return plan
 
 
+@deprecated("24.9", "25.3", addendum="Unused.")
 def _handle_menuinst(unlink_dists, link_dists):  # pragma: no cover
     # Always link/unlink menuinst first/last in case a subsequent
     # package tries to import it to create/remove a shortcut
@@ -466,6 +424,7 @@ def _handle_menuinst(unlink_dists, link_dists):  # pragma: no cover
     return unlink_dists, link_dists
 
 
+@deprecated("24.9", "25.3", addendum="Unused.")
 @time_recorder("install_actions")
 def install_actions(
     prefix,
@@ -532,6 +491,7 @@ def install_actions(
         return actions
 
 
+@deprecated("24.9", "25.3", addendum="Unused.")
 def get_blank_actions(prefix):  # pragma: no cover
     from collections import defaultdict
 
@@ -638,4 +598,6 @@ if __name__ == "__main__":
     # for testing new revert_actions() only
     from pprint import pprint
 
-    pprint(dict(revert_actions(sys.prefix, int(sys.argv[1]))))
+    deprecated.topic("24.9", "25.3", topic="`conda.plan` as an entrypoint")
+
+    pprint(dict(_revert_actions(sys.prefix, int(sys.argv[1]))))
